+++ owrx/config/defaults.py
@@ -20,6 +20,7 @@ defaultConfig = PropertyLayer(
     digimodes_fft_size=2048,
     digital_voice_unvoiced_quality=1,
     digital_voice_dmr_id_lookup=True,
+    digital_voice_nxdn_id_lookup=True,
     sdrs=PropertyLayer(
         rtlsdr=PropertyLayer(
             name="RTL-SDR USB Stick",