+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 **1.1.x - unreleased**
 - Reworked most graphical elements as SVGs for faster loadtimes and crispier display on hi-dpi displays
 - Updated pipelines to match changes in digiham
-- Changed D-Star integration to use new decoder in digiham
-- Added D-Star metadata display
+- Changed D-Star and NXDN integrations to use new decoders from digiham
+- Added D-Star and NXDN metadata display
 
 **1.0.0**
 - Introduced `squelch_auto_margin` config option that allows configuring the auto squelch level